compatibility with Qt 4.5 (got it wrong, last time)
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@3648 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ retroshare-gui/src/gui/common/RSTreeWidgetItem.cpp
@@ -119,7 +119,7 @@ bool RSTreeWidgetItem::operator<(const QTreeWidgetItem &other) const
     case 0: //integer type
         return v1.toLongLong() < v2.toLongLong();
     case 1: //floating point
-        return v1.toFloat() < v2.toFloat();
+        return v1.toDouble() < v2.toDouble();
     default:
         return (v1.toString().compare (v2.toString(), Qt::CaseInsensitive) < 0);
     }
